Save test result immediately after openning ResultsActivity instead of letting the user chose to save.
We don't get any event if the application is being killed, for example from the recents screen. As such we would not save the result in that case. We could have saved the result when the screen is first shown and then delete it if the user choses not to have it saved.